+++ CHANGELOG.md
@@ -4,11 +4,14 @@
 
 All notable changes to [this project](README.md) will be documented in this
 file. This project adheres to [Semantic Versioning](http://semver.org).
- 
+
+### 3.1.2
+* removed SLF4J Dependency, error logging now uses a more flexible BiFunction
+* [MimeTypes](src/main/java/diergo/csv/MimeTypes.java) is now deprecated s it is not ore functionality
 
 ### 3.1.1
 * minimal Java 9 support ([#7](https://github.com/aburmeis/decs/issues/7))
-* built with Java 11
+* built with Java 8 and 11
 * [Appendables](src/main/java/diergo/csv/Appendables.java) are now supported instead of Writers ([#6](https://github.com/aburmeis/decs/issues/6))
 * migrated tests to [JUnit 5](https://junit.org/junit5/)
 * [Writers](src/main/java/diergo/csv/Writers.java) is deprecated and will be removed in a future release